requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.32.3
apscheduler==3.10.4
tinydb==4.8.0
pytz==2024.1
